libteec: allow passing settings to context initialization
Add a new function, TEEC_InitializeContext2, that allows a developer to
pass configuration parameters in the form of distinct settings. One of the
new settings is for configuring OCALLs, where the caller sets a callback
handler for when an OCALL arrives from a TA.

Having multiple settings in this manner helps reduce the number of
auxiliary functions necessary. Were new functionality to be added in the
future, no new functions would need to be introduced. Instead, one would
only require a new setting.

Signed-off-by: Hernan Gatta <hegatta@microsoft.com>